aion/services/event-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
asyncpg==0.29.0
paho-mqtt==2.1.0
pydantic==2.9.2
python-json-logger==2.0.7
